fix the fix for duplicate ids in burp extension
The previous fix set a unique id whenever the burp extension generated
id was encountered, but this lead to multiple instances of the same
finding not being grouped.

So now we set specific ids for common findings based on their name and
warn on the others so that we can add more in the future if needed.